nvim: Move remaining functionality over to the new lua/ethersync.lua
And add some default config in plugin/ethersync.lua.
